feat(hub-transit-route): 공통 파일 추가
- global 패키지 내 공통 클래스들 추가
