--- hub-transit-route-service/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'hub-transit-route-service'
