--- hub-transit-route-service/src/main/java/on/logistics/hubtransitrouteservice/global/configuration/JpaAuditingConfig.java
@@ -1,10 +1,0 @@
-package on.logistics.hubtransitrouteservice.global.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-
-}
