--- hub-transit-route-service/src/main/java/on/logistics/hubtransitrouteservice/global/configuration/PageableConfig.java
@@ -1,20 +1,0 @@
-package on.logistics.hubtransitrouteservice.global.configuration;
-
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import on.logistics.hubtransitrouteservice.global.resolver.PageableVerificationArgumentResolver;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class PageableConfig implements WebMvcConfigurer {
-
-    private final PageableVerificationArgumentResolver pageableVerificationArgumentResolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(pageableVerificationArgumentResolver);
-    }
-}
