--- hub-transit-route-service/src/main/java/on/logistics/hubtransitrouteservice/global/configuration/QuerydslConfig.java
@@ -1,19 +1,0 @@
-package on.logistics.hubtransitrouteservice.global.configuration;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class QuerydslConfig {
-
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-}
