--- hub-transit-route-service/src/main/java/on/logistics/hubtransitrouteservice/global/configuration/WebClientConfig.java
@@ -1,15 +1,0 @@
-package on.logistics.hubtransitrouteservice.global.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-
-    @Bean
-    public WebClient webClient() {
-        return WebClient.builder().build();
-    }
-
-}
