--- hub-transit-route-service/src/main/java/on/logistics/hubtransitrouteservice/global/exception/CustomException.java
@@ -1,14 +1,0 @@
-package on.logistics.hubtransitrouteservice.global.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-
-    private final ExceptionCode exception;
-
-    public CustomException(ExceptionCode e) {
-        super(e.getMessage());
-        this.exception = e;
-    }
-}
