--- hub-transit-route-service/src/main/java/on/logistics/hubtransitrouteservice/global/exception/ExceptionCode.java
@@ -1,12 +1,0 @@
-package on.logistics.hubtransitrouteservice.global.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionCode {
-
-    String name();
-
-    HttpStatus getHttpStatus();
-
-    String getMessage();
-}
