+++ hub-transit-service/src/main/java/on/logistics/hubtransitservice/HubTransitServiceApplication.java
@@ -1,0 +1,15 @@
+package on.logistics.hubtransitservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@EnableDiscoveryClient
+@SpringBootApplication
+public class HubTransitServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HubTransitServiceApplication.class, args);
+    }
+
+}
