+++ hub-transit-service/src/main/java/on/logistics/hubtransitservice/global/configuration/FeignClientConfig.java
@@ -1,0 +1,10 @@
+package on.logistics.hubtransitservice.global.configuration;
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableFeignClients(basePackages = "on.logistics.hubtransitservice.infrastructure.clients")
+public class FeignClientConfig {
+
+}
