+++ hub-transit-service/src/main/java/on/logistics/hubtransitservice/global/configuration/JpaAuditingConfig.java
@@ -1,0 +1,10 @@
+package on.logistics.hubtransitservice.global.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+
+}
