+++ hub-transit-service/src/main/java/on/logistics/hubtransitservice/global/enums/PageSortBy.java
@@ -1,0 +1,22 @@
+package on.logistics.hubtransitservice.global.enums;
+
+import java.util.Arrays;
+import lombok.Getter;
+
+@Getter
+public enum PageSortBy {
+    CREATED_AT("created_at"),
+    UPDATED_AT("updated_at"),
+    ID("id");
+
+    private final String sortBy;
+
+    PageSortBy(String sortBy) {
+        this.sortBy = sortBy;
+    }
+
+    public static boolean isValid(String sortBy) {
+        return Arrays.stream(values())
+            .anyMatch(e -> e.getSortBy().equals(sortBy));
+    }
+}
