+++ hub-transit-service/src/main/java/on/logistics/hubtransitservice/global/exception/CommonExceptionCode.java
@@ -1,0 +1,18 @@
+package on.logistics.hubtransitservice.global.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum CommonExceptionCode implements ExceptionCode {
+
+    ILLEGAL_ARGUMENT(HttpStatus.BAD_REQUEST, "잘못된 매개변수입니다."),
+    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버에서 문제가 발생되었습니다."),
+    PAGE_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 페이지를 찾을 수 없습니다."),
+    ;
+
+    private final HttpStatus httpStatus;
+    private final String message;
+}
