+++ hub-transit-service/src/main/java/on/logistics/hubtransitservice/global/exception/ExceptionCode.java
@@ -1,0 +1,12 @@
+package on.logistics.hubtransitservice.global.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ExceptionCode {
+
+    String name();
+
+    HttpStatus getHttpStatus();
+
+    String getMessage();
+}
