+++ hub-transit-service/src/test/java/on/logistics/hubtransitservice/HubTransitServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package on.logistics.hubtransitservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HubTransitServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
